refactor(provider): extract message conversion + token estimation
Split the I partition out of extension.ts:
- provider/messages.ts — convertMessage (vscode parts → wire format, incl.
  image normalization, MiMo tool-result flattening, reasoning echo) +
  normalizeMessages / trimOldImagesFromHistoryInPlace / hasMessagePayload
- provider/tokens.ts — messageText / estimateChatMessageTokenCount /
  partToTokenCount / partToText etc.

Behavior-preserving; compile + 291 tests + lint green.
